refactor: enhance grimoire history management by adding event handling functions and updating imports
